+++ src/graph_notebook/__init__.py
@@ -3,4 +3,4 @@ Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 """
 
-__version__ = '3.0.6'
+__version__ = '3.0.8'